requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-limiter==3.5.0
flask-cors==4.0.0
gunicorn==21.2.0
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0